refactor(help): add src and types folders
